// src/app/services/openapi/model/tag.ts
/**
 * Digicatch
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Tag { 
    readonly id: string;
    name: string;
    description: string;
    points: number;
    color: string;
    readonly deleted: boolean;
    readonly updatedAt: string;
    image: string;
}